Rescue NPC + Buff Reward Implemented, Weapon Tweak
Approach the NPC and press E to interact with it. It will spawn two
buffs that the player will walk over and press E to collect.

Current buff list is for a flat health increase and a percent fire rate
increase.

The player's weapon has been modified to better handle changes in its
firing speed.