Pause Menu, Scene Implementations, Wave Spawner, Bug Fixes
Implemented Scenes: Main Menu, Main Level, Arena and interconnected them

Implemented Wave Spawner: added code to Destructable, PlayerController,
SlowPlayer, and added a WaveEnemy Script. And of course added the wave
implementation to the spawner script.

Bug Fixes:
Enemy gun get angle was incorrect and only shot towards the bottom right
corner fixed it so now it aims at the player

The Ranger Enemy was teleporting around too much and wouldn't attack
continuously when not teleporting so changed it in the code so it did
not teleport when it attacks and when stationary it continuously fires.
Also changed the prefab so it teleports every 5 seconds and fires every
1, also the bullet shoots faster

BulletCollisionHandler: commented out the sound code until we have a
sound to put in

PoolManager: Added a reset to the pool manager for when you start the
level

Pause Menu: can pause game by hitting escape key, can then resume game
or exit back to the main menu

Also baked all the navmeshs for arena and main level, and changed the
build settings to include all of the necessary scenes